Extracted filter code to a module
